# public/docs/api.yaml
openapi: 3.0.0
info:
  title: Little Home Mobile API - Wali Murid
  description: REST API endpoint untuk aplikasi mobile (Role 5 / Wali Murid) pada sistem Little Home.
  version: 1.0.0
servers:
  - url: http://localhost/little_home_api
    description: Local Server Utama
...
paths:
...
  /walimurid/pencatatan:
    get:
      tags:
        - Monitoring
      summary: List Riwayat Pencatatan (Perkembangan)
      description: Menarik rekaman keseharian (Rencana Pembelajaran & Observasi Anak) yang ditulis oleh guru.
      parameters:
        - in: query
          name: id_anak
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Data list pencatatan
...
  /walimurid/testimoni:
    get:
      tags:
        - Layanan Umum
      summary: Lihat List Testimoni Public
      responses:
        '200':
          description: Array of Object Testimoni
    post:
      tags:
        - Layanan Umum
      summary: Kirim Feedback Testimoni
      description: Mengunggah tanggapan (Membutuhkan Approval / Validation "status=0" pada database awal)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - nama
                - testimoni
              properties:
                nama:
                  type: string
                jabatan:
                  type: string
                  description: Optional (Default akan disetel "Wali Murid")
                testimoni:
                  type: string
      responses:
        '200':
          description: Testimoni Disimpan
        '400':
          description: Parameter Tidak Lengkap/Kosong
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []